+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 ### Added
 - The *scoreboard.yml* file defines the scoreboards to be displayed for each arena.
 - The `commands-at-end` option lets you execute commands at the end of a game.
+- The hook into Vault allows choosing how much money players should receive.
 
 ### Fixed
 - When there were two players left, the one who failed won the game.